Moved GetNameSize into a separate service
This will be removed eventiallu, but we need to firstclear up the old Device and Forns.Init things and that means removing the IPlatformServices interface.
