+++ src/Compatibility/Core/src/GTK/GtkPlatformServices.cs
@@ -32,33 +32,6 @@ namespace Microsoft.Maui.Controls.Compatibility.Platform.GTK
 			return AppDomain.CurrentDomain.GetAssemblies();
 		}
 
-		public double GetNamedSize(NamedSize size, Type targetElementType, bool useOldSizes)
-		{
-			switch (size)
-			{
-				case NamedSize.Default:
-					return 11;
-				case NamedSize.Micro:
-				case NamedSize.Caption:
-					return 12;
-				case NamedSize.Medium:
-					return 17;
-				case NamedSize.Large:
-					return 22;
-				case NamedSize.Small:
-				case NamedSize.Body:
-					return 14;
-				case NamedSize.Header:
-					return 46;
-				case NamedSize.Subtitle:
-					return 20;
-				case NamedSize.Title:
-					return 24;
-				default:
-					throw new ArgumentOutOfRangeException(nameof(size));
-			}
-		}
-
 		public async Task<Stream> GetStreamAsync(Uri uri, CancellationToken cancellationToken)
 		{
 			using (var client = new HttpClient())
