+++ src/Controls/src/Core/Device.cs
@@ -133,10 +133,9 @@ namespace Microsoft.Maui.Controls
 
 		/// <include file="../../docs/Microsoft.Maui.Controls/Device.xml" path="//Member[@MemberName='GetNamedSize'][2]/Docs" />
 		[EditorBrowsable(EditorBrowsableState.Never)]
-		public static double GetNamedSize(NamedSize size, Type targetElementType, bool useOldSizes)
-		{
-			return PlatformServices.GetNamedSize(size, targetElementType, useOldSizes);
-		}
+		public static double GetNamedSize(NamedSize size, Type targetElementType, bool useOldSizes) =>
+			DependencyService.Get<IFontNamedSizeService>()?.GetNamedSize(size, targetElementType, useOldSizes) ??
+			throw new NotImplementedException("The current platform does not implement the IFontNamedSizeService dependency service.");
 
 		public static class Styles
 		{
