+++ src/Controls/src/Core/IFontNamedSizeService.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Microsoft.Maui.Controls.Internals
+{
+	public interface IFontNamedSizeService
+	{
+		double GetNamedSize(NamedSize size, Type targetElementType, bool useOldSizes);
+	}
+}
